services/modelrna/src/requirements.txt
aioredis==2.0.0
amqp==5.0.9
anyio==3.3.3
asgiref==3.4.1
async-timeout==3.0.1
billiard==3.6.4.0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
fastapi==0.70.0
flatbuffers==2.0
h11==0.12.0
httptools==0.2.0
idna==3.2
kombu==5.2.3
numpy==1.21.2
onnxruntime==1.8.1
prompt-toolkit==3.0.20
protobuf==3.18.1
pycryptodomex==3.11.0
pydantic==1.8.2
PyJWT==2.2.0
python-dotenv==0.19.1
python-multipart==0.0.5
pytz==2021.3
PyYAML==5.4.1
redis==3.5.3
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
typing-extensions==3.10.0.2
uvicorn==0.15.0
uvloop==0.16.0
vine==5.0.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.0
